+++ changelog.md
@@ -1,5 +1,5 @@
 ### 2.6.0 (2023-12-xx)
-- added sickoLooper5 module
+- added sickoLooper3 and sickoLooper5 modules
 
 ### 2.5.10 (2023-11-19)
 - all player and sampler modules: added 'Store Sample in Patch' feature